filesystem: Linux again diverges from the pack.
As we are aiming to suport the other OS, we will treat
Linux as the exception as with memory use.

Linux uses /proc/mounts for all this however BSD/macOS use a
single struct and system call. The argument might be the
expense of the syscall, but for our purposes and sanity use
the greatest common demoninator again. Sorry Linux...

Seems procfs isn't going to scale very well in the future, am
not looking forward to what they use next...