openbsd: amd64.
Firstly, we don't need so many locks as we blocking the main loop.

On amd64 OpenBSD 6.8. Static are stored in tbss, so I know TLS was
releatively recently added properly, that's one issue.

I can only speculate, it's thread related and how mutexes are
handled on amd64? As well as having too many threads, also too
much blocking of the main loop...

Other changes, get rid of almost all sense of a "main" window...

Lots of stuff learnt through hacking...My folks are downstairs and
I can hear shaging on the TV...Christmas time...it's mean to be...
Sheesh